Move Package.js to adapter/Package.js
